+++ R/processing.R
@@ -12,12 +12,11 @@ runfolds <- function(j = NULL,
                      i = NULL,
                      formula = NULL,
                      model_args = NULL,
-                     par_cl = NULL,
-                     par_mode = NULL,
                      model_fun = NULL,
                      pred_fun = NULL,
                      imp_variables = NULL,
                      imp_permutations = NULL,
+                     imp_sample_from = "test",
                      err_fun = NULL,
                      train_fun = NULL,
                      importance = NULL,
@@ -29,7 +28,6 @@ runfolds <- function(j = NULL,
                      pooled_pred_train = NULL,
                      response = NULL,
                      progress = NULL,
-                     is_factor_prediction = NULL,
                      pooled_pred_test = NULL,
                      coords = NULL,
                      test_fun = NULL,
@@ -45,14 +43,16 @@ runfolds <- function(j = NULL,
   fold_number <- NULL
   repetition_number <- NULL
 
+  # predictor variables:
+  predvars <- all.vars(formula)[-1]
+
   # append fid to data to enable tracking of observations
-  data$fid <- seq(1:nrow(data)) # nolint
+  data$fid <- seq_len(nrow(data)) # nolint
 
   # Create training sample:
   nd_train <- data[current_sample[[j]]$train, ]
   if (!is.null(train_fun)) {
-    nd_train_sel <- train_fun(data = nd_train, param = train_param)
-    nd_train <- nd_train[nd_train_sel, ]
+    nd_train <- train_fun(data = nd_train, param = train_param)
     current_sample[[j]]$train <- nd_train$fid
   }
 
@@ -95,7 +95,8 @@ runfolds <- function(j = NULL,
       pooled_pred_test = NA,
       current_res = NA,
       current_impo = NA,
-      not_converged_folds = not_converged_folds
+      not_converged_folds = not_converged_folds,
+      is_factor_prediction = NA
     ))
   }
 
@@ -107,6 +108,8 @@ runfolds <- function(j = NULL,
     pred_train <- do.call(pred_fun, args = pargs)
   }
   rm(pargs)
+
+  is_factor_prediction <- is.factor(pred_train)
 
   # Calculate error measures on training sample:
 
@@ -139,7 +142,8 @@ runfolds <- function(j = NULL,
       current_res = NA,
       current_impo = NA,
       not_converged_folds = not_converged_folds,
-      resampling = current_sample
+      resampling = current_sample,
+      is_factor_prediction = is_factor_prediction
     ))
   }
 
@@ -147,8 +151,7 @@ runfolds <- function(j = NULL,
   pooled_pred_train <- c(pooled_pred_train, pred_train)
 
   if (!is.null(test_fun)) {
-    nd_test_sel <- test_fun(data = nd_test, param = test_param)
-    nd_test <- nd_test[nd_test_sel, ]
+    nd_test <- test_fun(data = nd_test, param = test_param)
     current_sample[[j]]$test <- nd_test$fid
   }
 
@@ -162,8 +165,16 @@ runfolds <- function(j = NULL,
     nd_test <- remove_missing_levels(fit, nd_test) # nolint
   }
 
-  # remove NAs in data.frame if levels are missing
-  nd_test <- na.omit(nd_test)
+  # remove NAs in test data.frame
+  if (length(predvars) > 0) {
+    if (length(predvars) == 1) {
+      omit_na <- !is.na(nd_test[, predvars])
+    } else {
+      omit_na <- apply(nd_test[, predvars], 1, function(x) !any(is.na(x)))
+    }
+    nd_test <- nd_test[omit_na, ]
+    ##nd_test <- na.omit(nd_test)
+  }
 
   # Apply model to test sample:
   pargs <- c(list(object = fit, newdata = nd_test), pred_args)
@@ -202,8 +213,6 @@ runfolds <- function(j = NULL,
       repetition_number
     ))
 
-    is_factor_prediction <<- is.factor(pred_test)
-
     not_converged_folds <- not_converged_folds + 1
 
     return(list(
@@ -214,14 +223,13 @@ runfolds <- function(j = NULL,
       current_res = NA,
       current_impo = NA,
       not_converged_folds = not_converged_folds,
-      resampling = current_sample
+      resampling = current_sample,
+      is_factor_prediction = is_factor_prediction
     ))
   }
 
   pooled_obs_test <- c(pooled_obs_test, nd_test[, response])
   pooled_pred_test <- c(pooled_pred_test, pred_test)
-
-  is_factor_prediction <<- is.factor(pred_test)
 
   ### Permutation-based variable importance assessment:
   if (importance == TRUE) {
@@ -234,12 +242,23 @@ runfolds <- function(j = NULL,
       nd_test <- remove_missing_levels(fit, nd_test) # nolint
     }
 
-    # remove NAs in data.frame if levels are missing
-    nd_test <- na.omit(nd_test)
+    # remove NAs in test data.frame
+    if (length(predvars) > 0) {
+      if (length(predvars) == 1) {
+        omit_na <- !is.na(nd_test[, predvars])
+      } else {
+        omit_na <- apply(nd_test[, predvars], 1, function(x) !any(is.na(x)))
+      }
+      nd_test <- nd_test[omit_na, ]
+      ##nd_test <- na.omit(nd_test)
+    }
+
+    # Backup copy for permutation:
+    nd_test_bak <- nd_test
 
     # does this ever happen??
     # nocov start
-    if (is.null(current_res[[j]]$test)) {
+    if (is.null(current_res[[j]]$test) & (imp_sample_from == "test")) {
       current_impo[[j]] <- c()
       if (!progress == FALSE) {
         cat(date(), "-- skipping variable importance\n")
@@ -250,28 +269,39 @@ runfolds <- function(j = NULL,
       }
       imp_temp <- imp_one_rep
 
-      # Parallelize this: ???
-      for (cnt in 1:imp_permutations) {
+      for (cnt in seq_len(imp_permutations)) {
         # Some output on screen:
         if (progress == "all" | progress == TRUE & (cnt > 1)) {
           if (log10(cnt) == floor(log10(cnt))) {
             cat(
               date(), "Repetition", i, "- Fold", j,
-              "- permutation-count:", cnt, "\n"
+              "- permutation count:", cnt, "\n"
             )
           }
         }
+
+        # Permutation data:
+        if (imp_sample_from == "test") {
+          nd_permute <- nd_test_bak
+        } else if (imp_sample_from == "train") {
+          nd_permute <- nd_train
+        } else if (imp_sample_from == "all") {
+          nd_permute <- data
+        }
         # Permutation indices:
-        permut <- sample(seq_len(nrow(nd_test)), replace = FALSE)
+        replace <- (nrow(nd_permute) < nrow(nd_test))
+        permut <- sample(seq_len(nrow(nd_permute)), replace = replace, size = nrow(nd_test))
+        # Permutation data:
+        nd_permute <- nd_permute[permut,]
 
         # For each variable:
         for (vnm in imp_variables) {
 
           # Get undisturbed backup copy of test sample:
-          nd_test <- nd_bak
+          nd_test <- nd_test_bak
 
-          # Permute variable vnm:
-          nd_test[, vnm] <- nd_test[, vnm][permut]
+          # Get permuted variable vnm:
+          nd_test[, vnm] <- nd_permute[, vnm]
           # Apply model to perturbed test sample:
           pargs <- c(list(object = fit, newdata = nd_test), pred_args)
           if (is.null(pred_fun)) {
@@ -309,7 +339,8 @@ runfolds <- function(j = NULL,
     current_res = current_res,
     current_impo = current_impo,
     not_converged_folds = not_converged_folds,
-    resampling = current_sample
+    resampling = current_sample,
+    is_factor_prediction = is_factor_prediction
   ))
 }
 
@@ -317,6 +348,7 @@ runfolds <- function(j = NULL,
 #' @description Runs model fitting, error estimation and variable importance
 #' on fold level
 #' @keywords internal
+#' @importFrom future.apply future_lapply
 #' @export
 #'
 
@@ -325,31 +357,30 @@ runreps <- function(current_sample = NULL,
                     data = NULL,
                     formula = NULL,
                     model_args = NULL,
-                    par_cl = NULL,
                     do_gc = NULL,
                     imp_one_rep = NULL,
                     model_fun = NULL,
                     pred_fun = NULL,
                     imp_variables = NULL,
                     imp_permutations = NULL,
+                    imp_sample_from = "test",
                     err_fun = NULL,
                     importance = NULL,
                     current_res = NULL,
                     current_impo = NULL,
                     pred_args = NULL,
                     progress = NULL,
+                    mode_fold = "sequential",
                     pooled_obs_train = NULL,
                     pooled_obs_test = NULL,
                     pooled_pred_train = NULL,
                     response = NULL,
-                    is_factor_prediction = NULL,
                     pooled_pred_test = NULL,
                     test_fun = NULL,
                     test_param = NULL,
                     train_fun = NULL,
                     train_param = NULL,
                     coords = NULL,
-                    par_mode = NULL,
                     i = NULL) {
 
   # output data structures
@@ -372,35 +403,103 @@ runreps <- function(current_sample = NULL,
     cat(date(), "Repetition", i, "\n") # nocov
   }
 
-  lapply(seq_along(current_sample), function(rep) {
-    runfolds(
-      j = rep,
-      data = data,
-      current_sample = current_sample,
-      formula = formula,
-      par_mode = par_mode,
-      pred_fun = pred_fun,
-      model_args = model_args,
-      model_fun = model_fun,
-      imp_permutations = imp_permutations,
-      imp_variables = imp_variables,
-      is_factor_prediction = is_factor_prediction,
-      importance = importance,
-      current_res = current_res,
-      pred_args = pred_args,
-      response = response,
-      par_cl = par_cl,
-      coords = coords,
-      progress = progress,
-      pooled_obs_train = pooled_obs_train,
-      pooled_obs_test = pooled_obs_test,
-      test_fun = test_fun,
-      test_param = test_param,
-      train_fun = train_fun,
-      train_param = train_param,
-      err_fun = err_fun
-    )
-  }) -> runfolds_list
+  if (mode_fold == "sequential") {
+    lapply(seq_along(current_sample), function(rep) {
+      runfolds(
+        j = rep,
+        data = data,
+        current_sample = current_sample,
+        formula = formula,
+        pred_fun = pred_fun,
+        model_args = model_args,
+        model_fun = model_fun,
+        imp_permutations = imp_permutations,
+        imp_variables = imp_variables,
+        imp_sample_from = imp_sample_from,
+        importance = importance,
+        current_res = current_res,
+        pred_args = pred_args,
+        response = response,
+        coords = coords,
+        progress = progress,
+        pooled_obs_train = pooled_obs_train,
+        pooled_obs_test = pooled_obs_test,
+        test_fun = test_fun,
+        test_param = test_param,
+        train_fun = train_fun,
+        train_param = train_param,
+        err_fun = err_fun
+      )
+    }) -> runfolds_list
+  } else if (mode_fold == "future") {
+    future.apply::future_lapply(seq_along(current_sample), function(rep) {
+      runfolds(
+        j = rep,
+        data = data,
+        current_sample = current_sample,
+        formula = formula,
+        pred_fun = pred_fun,
+        model_args = model_args,
+        model_fun = model_fun,
+        imp_permutations = imp_permutations,
+        imp_variables = imp_variables,
+        imp_sample_from = imp_sample_from,
+        importance = importance,
+        current_res = current_res,
+        pred_args = pred_args,
+        response = response,
+        coords = coords,
+        progress = progress,
+        pooled_obs_train = pooled_obs_train,
+        pooled_obs_test = pooled_obs_test,
+        test_fun = test_fun,
+        test_param = test_param,
+        train_fun = train_fun,
+        train_param = train_param,
+        err_fun = err_fun
+      )
+    },
+    future.seed = TRUE
+    ) -> runfolds_list
+  } else if (mode_fold == "loop") {
+    runfolds_list <- list()
+    for (i_fold in seq_along(current_sample)) {
+      runfolds_list[[i_fold]] <-
+        runfolds(
+          j = i_fold,
+          data = data,
+          current_sample = current_sample,
+          formula = formula,
+          pred_fun = pred_fun,
+          model_args = model_args,
+          model_fun = model_fun,
+          imp_permutations = imp_permutations,
+          imp_variables = imp_variables,
+          imp_sample_from = imp_sample_from,
+          importance = importance,
+          current_res = current_res,
+          pred_args = pred_args,
+          response = response,
+          coords = coords,
+          progress = progress,
+          pooled_obs_train = pooled_obs_train,
+          pooled_obs_test = pooled_obs_test,
+          test_fun = test_fun,
+          test_param = test_param,
+          train_fun = train_fun,
+          train_param = train_param,
+          err_fun = err_fun
+        )
+    }
+  }
+
+  is_factor_prediction <- sapply(runfolds_list, function(x) x$is_factor_prediction)
+  is_factor_prediction <- is_factor_prediction[ !is.na(is_factor_prediction) ]
+  if (length(is_factor_prediction) == 0) {
+    is_factor_prediction <- FALSE
+  } else {
+    is_factor_prediction <- any(is_factor_prediction) # or all?
+  }
 
   # merge sublists of each fold into one list
   # http://stackoverflow.com/questions/32557131/adding-a-vector-to-each-sublist-within-a-list-r # nolint
@@ -410,7 +509,9 @@ runreps <- function(current_sample = NULL,
   if (all(is.na(runfolds_merged$pooled_obs_train))) {
     return(list(
       error = NA,
-      pooled_error = NA, importance = NA
+      pooled_error = NA, importance = NA,
+      not_converged_folds = runfolds_merged$not_converged_folds,
+      resampling = runfolds_merged$resampling
     ))
   }
 
@@ -438,7 +539,7 @@ runreps <- function(current_sample = NULL,
       levels = lev
     )
 
-    if (is_factor_prediction == TRUE) {
+    if (is_factor_prediction) {
       pooled_only$pooled_pred_train <- factor(lev[
         pooled_only$pooled_pred_train
       ], levels = lev)