+++ R/summary_functions.R
@@ -13,8 +13,8 @@
 #'
 #' @param object `sperroresterror` resp. `sperrorestcombinederror` error object
 #'   calculated by [sperrorest]
-#' @param level Level at which errors are summarized: 0: overall; 1: repetition;
-#'   2: fold
+#' @param level Level at which errors are summarized: 0: overall (i.e. across
+#'   all repetitions); 1: repetition; 2: fold
 #' @param pooled If `TRUE` (default), mean and standard deviation etc are
 #'   calculated between fold-level error estimates. If `FALSE`, apply first a
 #'   [weighted.mean] among folds before calculating mean, standard deviation etc
@@ -44,36 +44,76 @@
 #' @seealso [sperrorest]
 #'
 #' @export
+#' @importFrom dplyr bind_rows
 summary.sperroresterror <- function(object, # nolint start
-                                    level = 0,
-                                    pooled = TRUE,
-                                    na.rm = TRUE,
-                                    ...) {
+                                                 level = 0,
+                                                 pooled = TRUE,
+                                                 na.rm = TRUE,
+                                                 ...) {
   err <- unclass(object)
+
+  # Check if object really contains error estimates:
+  is_err_object <- function(x) {
+    res <- FALSE
+    if (is.list(x)) {
+      if (all(c("train", "test") %in% names(x))) {
+        if (is.list(x$train)) {
+          res <- TRUE
+        }
+      }
+    }
+    res
+  }
+
+  get_dist <- function(y) {
+    ifelse(any(names(y) == "distance"), y$distance, -1)
+  }
+
+  get_err <- function(y) {
+    if (!is_err_object(y)) return(NULL)
+    res <- data.frame(
+      train = y$train,
+      test = y$test,
+      distance = get_dist(y)
+    )
+    if (length(y$train) == 1) {
+      colnames(res) <- c(
+        paste0("train.", names(y$train)),
+        paste0("test.", names(y$test)),
+        "distance"
+      )
+    }
+    res
+  }
+
+  if (level <= 2) {
+    for (i in seq_along(err)) {
+      err_rep <- list()
+      cnt <- 0
+      for (j in seq_along(err[[i]])) {
+        tmp <- get_err(err[[i]][[j]])
+        if (!is.null(tmp)) {
+          cnt <- cnt + 1
+          err_rep[[cnt]] <- tmp
+        }
+      }
+      err[[i]] <- as.data.frame(dplyr::bind_rows(err_rep))
+    }
+  }
+
   # nolint end
   if (pooled) {
-    if (level <= 2) {
-      err <- lapply(err, function(x) {
-        t(sapply(x, function(y) {
-          data.frame(
-            train = y$train,
-            test = y$test,
-            distance = ifelse(any(names(y) == "distance"), y$distance,
-              -1
-            )
-          )
-        }))
-      })
-    }
     if (level <= 1) {
       errdf <- err[[1]]
       if (length(err) > 1) {
         for (i in 2:length(err)) {
-          errdf <- rbind(errdf, err[[i]])
+          if (nrow(err[[i]] > 0))
+            errdf <- rbind(errdf, err[[i]])
         }
       }
       rownames(errdf) <- NULL
       err <- as.data.frame(errdf)
+      rm(errdf)
     }
     if (level <= 0) {
       err <- data.frame(
@@ -90,43 +130,27 @@ summary.sperroresterror <- function(object, # nolint start
       )
     }
   } else {
-    if (level <= 2) {
-      err <- lapply(err, function(x) {
-        t(sapply(x, function(y) {
-          data.frame(
-            train = y$train,
-            test = y$test,
-            distance = ifelse(any(names(y) == "distance"), y$distance,
-              -1
-            )
-          )
-        }))
-      })
-    }
     if (level <= 1) {
       ### w = summary.partition(resampling) ?????
       err <- lapply(err, function(x) {
         apply(x, 2, function(y) {
-          weighted.mean(unlist(y),
-            na.rm = na.rm
-          )
+          weighted.mean(unlist(y), na.rm = na.rm)
         })
       })
-      nms <- names(err)
-      err <- as.data.frame(t(as.data.frame(err)))
-      rownames(err) <- nms
+      sel <- sapply(err, length) > 0
+      err <- err[sel]
+      err <- as.data.frame(dplyr::bind_rows(err))
     }
     if (level <= 0) {
       err <- data.frame(
-        mean = sapply(err, mean), sd = sapply(err, sd),
-        median = sapply(
-          err,
-          median
-        ), IQR = sapply(err, IQR)
+        mean = sapply(err, mean),
+        sd = sapply(err, sd),
+        median = sapply(err, median),
+        IQR = sapply(err, IQR)
       )
     }
   }
-  return(err)
+  err
 }
 
 #' @rdname summary.sperrorest
@@ -142,12 +166,13 @@ summary.sperrorestreperror <- function(object, # nolint
   object <- as.data.frame(object)
   if (level <= 0) {
     object <- data.frame(
-      mean = sapply(object, mean), sd = sapply(object, sd),
-      median = sapply(object, median),
-      IQR = sapply(object, IQR)
+      mean = sapply(object, mean, na.rm = na.rm),
+      sd = sapply(object, sd, na.rm = na.rm),
+      median = sapply(object, median, na.rm = na.rm),
+      IQR = sapply(object, IQR, na.rm = na.rm)
     )
   }
-  return(object)
+  object
 }
 
 #' @title Summarize variable importance statistics obtained by {sperrorest}
@@ -162,7 +187,7 @@ summary.sperrorestreperror <- function(object, # nolint
 #'   with argument `importance = TRUE`
 #' @inheritParams summary.sperroresterror
 #' @param which optional character vector specifying selected variables for
-#'   which the importances should be summarized (to do: check implementation)
+#'   which the importances should be summarized
 #'
 #' @return a list or data.frame, depending on the `level` of aggregation
 #'
@@ -173,34 +198,49 @@ summary.sperrorestimportance <- function(object, # nolint start
                                          which = NULL,
                                          ...) {
   # nolint end
-  arrdim <- c(length(object), length(object[[1]]), dim(object[[1]][[1]]))
+  # Create array that will hold the variable importance results:
+  nfolds <- sapply(object, length)
+  if (any(nfolds != nfolds[1]) & !na.rm) {
+    warning("Repetitions have a varying number of folds, therefore NAs will be produced; chaging 'na.rm' to TRUE.")
+    na.rm <- TRUE
+  }
+  arrdim <- c(length(object), max(nfolds), dim(object[[1]][[1]]))
   arrdimnames <- list(
-    names(object), names(object[[1]]),
-    rownames(object[[1]][[1]]), colnames(object[[1]][[1]])
+    names(object), # names of repetitions
+    names(object[[which.max(nfolds)]]), # names of folds
+    rownames(object[[1]][[1]]), # names of
+    colnames(object[[1]][[1]])
   )
   arr <- array(NA, dim = arrdim, dimnames = arrdimnames)
+  # Fill the array with the results:
   for (i in seq_along(object)) {
-    for (j in seq_along(object[[i]])) {
-      arr[i, j, , ] <- as.matrix(object[[i]][[j]])
+    if (nfolds[i] >= 1) {
+      for (j in seq_along(object[[i]])) {
+        if (is.data.frame(object[[i]][[j]])) {
+          arr[i, j, , ] <- as.matrix(object[[i]][[j]])
+        } else {
+          warning("ignoring unexpected object in repetition ", i, ", fold ", j)
+        }
+      }
     }
   }
+  # Summarize at the repetition level:
   if (level <= 1) {
     arr <- apply(arr, c(1, 3, 4), mean, na.rm = na.rm)
   }
+  # Summarize at the overall level, i.e. across all repetitions:
   if (level <= 0) {
     if (is.null(which)) {
       arr <- data.frame(
         mean = apply(arr, c(2, 3), mean, na.rm = na.rm),
         sd = apply(arr, c(2, 3), sd, na.rm = na.rm),
-        median = apply(arr, c(2, 3), median,
-          na.rm = na.rm
-        ),
+        median = apply(arr, c(2, 3), median, na.rm = na.rm),
         IQR = apply(arr, c(2, 3), IQR, na.rm = na.rm)
       )
     }
-    # when does this happen?
     else {
-      arr <- arr[, , which] # nocov start
+      # summarize importance of selected features:
+      arr <- arr[, which, ] # nocov start
       arr <- data.frame(
         mean = apply(arr, 2, mean, na.rm = na.rm),
         sd = apply(arr, 2, sd, na.rm = na.rm),
@@ -209,7 +249,7 @@ summary.sperrorestimportance <- function(object, # nolint start
       ) # nocov end
     }
   }
-  return(arr)
+  arr
 }